Extracted.
